object detection/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=1.0.0
